backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
Pillow==10.4.0
open_clip_torch==2.26.1
torch==2.4.0+cpu
torchvision==0.19.0+cpu
pydantic==2.9.2

--extra-index-url https://download.pytorch.org/whl/cpu
